+++ apis/python/src/tiledbsoma/soma_group.cc
@@ -50,7 +50,6 @@ void load_soma_group(py::module& m) {
             [](SOMAGroup& group) {
                 return group.mode() == OpenMode::read ? "r" : "w";
             })
-        .def("reopen", &SOMAGroup::reopen)
         .def("close", &SOMAGroup::close)
         .def_property_readonly(
             "closed",