+++ libtiledbsoma/src/soma/soma_group.cc
@@ -150,11 +150,6 @@ void SOMAGroup::open(
     fill_caches();
 }
 
-std::unique_ptr<SOMAGroup> SOMAGroup::reopen(
-    OpenMode mode, std::optional<TimestampRange> timestamp) {
-    return std::make_unique<SOMAGroup>(mode, uri_, ctx_, name_, timestamp);
-}
-
 void SOMAGroup::close() {
     if (group_->query_type() == TILEDB_WRITE)
         cache_group_->close();