+++ libtiledbsoma/src/soma/soma_group.h
@@ -107,16 +107,6 @@ class SOMAGroup : public SOMAObject {
         OpenMode mode, std::optional<TimestampRange> timestamp = std::nullopt);
 
     /**
-     * Return a new SOMAGroup with the given mode at the current Unix timestamp.
-     *
-     * @param mode if the OpenMode is not given, If the SOMAObject was opened in
-     * READ mode, reopen it in WRITE mode and vice versa
-     * @param timestamp Timestamp
-     */
-    std::unique_ptr<SOMAGroup> reopen(
-        OpenMode mode, std::optional<TimestampRange> timestamp = std::nullopt);
-
-    /**
      * Close the SOMAGroup object.
      */
     void close();