+++ apis/python/HISTORY.md
@@ -25,6 +25,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - \[[#4071](https://github.com/single-cell-data/TileDB-SOMA/pull/4071)\] [python] A `tiledb_timestamp` with value of zero is now equivalent to an unspecified timestamp (or `None`), and will be a synonym for "current time". Prior to this fix, a zero-valued timestamp would generate errors or unpredictable results.
 - \[[#4103](https://github.com/single-cell-data/TileDB-SOMA/pull/4103)\] [python] Do not attempt to resize empty measurements in tiledbsoma.io.prepare_experiment. This fixes a bug where calling `prepare_experiment` would fail if the Experiment contained any empty Measurements.
 
+- \[[#4106](https://github.com/single-cell-data/TileDB-SOMA/pull/4106)\] [python] The `SOMAObject.reopen` method no longer modifies the original `SOMAObject`. It only creates a new instance of a `SOMAObject` in the requested mode and at the requested timestamp. Attempting to `reopen` an object with unwritten metadata now raises a `SOMAError`.
+
 ### Security
 
 ## [Release 1.17.0]